# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Banking System",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/v1/transactions": {
      "post": {
        "tags": ["transaction"],
        "summary": "Create Transaction",
        "description": "Create a new transaction.",
        "operationId": "createTransaction",
        "requestBody": {
          "description": "Transaction data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sourceAccountId": {
                    "type": "integer"
                  },
                  "destinationAccountId": {
                    "type": "integer"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transaction": {
                          "$ref": "#/components/schemas/Transaction"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["transaction"],
        "summary": "Get Transaction",
        "description": "Get all transaction.",
        "operationId": "getTransaction",
        "responses": {
          "201": {
            "description": "Transaction successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "transaction": {
                          "$ref": "#/components/schemas/Transaction"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "sourceAccountId": {
            "type": "integer"
          },
          "destinationAccountId": {
            "type": "integer"
          },
          "amount": {
            "type": "number"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Type token without 'Bearer'",
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}